refactor/25: Unified commands to `configure-repo`
